fix: model refactor and representation
