+++ AssrWeb/MetaCommon/models/__init__.py
@@ -1,0 +1,6 @@
+from .datafile import DataFile
+from .metadata import Metadata
+from .tags import Tags
+
+from .test_models import (TestingTagsModel, 
+    TestingMetadataModel, TestingDataFile)
