+++ AssrWeb/MetaCommon/models/metadata.py
@@ -1,0 +1,27 @@
+from django.db import models
+from django.core import validators
+from .tags import Tags
+
+
+class Metadata(models.Model):
+    name = models.CharField(
+        max_length=50,
+        validators=[validators.MaxLengthValidator(50)]
+    )
+    author = models.CharField(
+        max_length=50,
+        validators=[validators.MaxLengthValidator(50)]
+    )
+    creationData = models.DateTimeField(
+        auto_now_add=True,
+        blank=True,
+        verbose_name='Дата создания'
+    )
+    keyValue = models.JSONField()
+    tag = models.ManyToManyField(Tags)
+
+    class Meta:
+        abstract = True
+
+    def __str__(self):
+        return self.name
