+++ AssrWeb/MetaCommon/models/test_models/__init__.py
@@ -1,0 +1,3 @@
+from .tags import TestingTagsModel
+from .metadata import TestingMetadataModel
+from .datafile import TestingDataFile
