+++ AssrWeb/MetaCommon/models/test_models/metadata.py
@@ -1,0 +1,8 @@
+from django.db import models
+from MetaCommon.models import Metadata
+from .tags import TestingTagsModel
+
+
+class TestingMetadataModel(Metadata):
+    objects = models.Manager()
+    tag = models.ManyToManyField(TestingTagsModel)
