+++ AssrWeb/MetaCommon/models/test_models/tags.py
@@ -1,0 +1,6 @@
+from django.db import models
+from MetaCommon.models import Tags
+
+
+class TestingTagsModel(Tags):
+    objects = models.Manager()
