+++ AssrWeb/UploadSource/models/__init__.py
@@ -1,2 +1,3 @@
-from .source_file import SourceFile, SourceMetadata
+from .source_file import SourceFile
+from .source_file_meta import SourceMetadata
 from .source_file_tag import SourceTags
